fix: Fix the bug in debugging Tag (#2250)
* fix: 修复调试tag不显示bug

* perf

* perf: 优化代码

* fix: 返回新对象
