+++ projects/app/src/global/core/chat/utils.ts
@@ -1,0 +1,43 @@
+import { ChatRoleEnum } from '@fastgpt/global/core/chat/constants';
+import { ChatHistoryItemResType, ChatItemType } from '@fastgpt/global/core/chat/type';
+import { SearchDataResponseItemType } from '@fastgpt/global/core/dataset/type';
+import { FlowNodeTypeEnum } from '@fastgpt/global/core/workflow/node/constant';
+
+const isLLMNode = (item: ChatHistoryItemResType) =>
+  item.moduleType === FlowNodeTypeEnum.chatNode || item.moduleType === FlowNodeTypeEnum.tools;
+
+export function transformPreviewHistories(histories: ChatItemType[]) {
+  return histories.map((item) => {
+    return {
+      ...addStatisticalDataToHistoryItem(item),
+      responseData: undefined
+    };
+  });
+}
+
+export function addStatisticalDataToHistoryItem(historyItem: ChatItemType) {
+  if (historyItem.obj !== ChatRoleEnum.AI) return historyItem;
+  if (historyItem.totalQuoteList !== undefined) return historyItem;
+  const flatResData: ChatHistoryItemResType[] =
+    historyItem.responseData
+      ?.map((item) => {
+        if (item.pluginDetail || item.toolDetail) {
+          return [item, ...(item.pluginDetail || []), ...(item.toolDetail || [])];
+        }
+        return item;
+      })
+      .flat() || [];
+  return {
+    ...historyItem,
+    llmModuleAccount: flatResData.filter(isLLMNode).length,
+    totalQuoteList: flatResData
+      .filter((item) => item.moduleType === FlowNodeTypeEnum.datasetSearchNode)
+      .map((item) => item.quoteList)
+      .flat()
+      .filter(Boolean) as SearchDataResponseItemType[],
+    totalRunningTime: Number(
+      flatResData.reduce((sum, item) => sum + (item.runningTime || 0), 0).toFixed(2)
+    ),
+    historyPreviewLength: flatResData.find(isLLMNode)?.historyPreview?.length
+  };
+}
