+++ projects/app/src/pages/api/core/chat/init.ts
@@ -12,7 +12,7 @@ import { getAppLatestVersion } from '@fastgpt/service/core/app/controller';
 import { NextAPI } from '@/service/middleware/entry';
 import { ReadPermissionVal } from '@fastgpt/global/support/permission/constant';
 import { FlowNodeTypeEnum } from '@fastgpt/global/core/workflow/node/constant';
-import transformPreviewHistories from '@/service/core/chat/utils';
+import { transformPreviewHistories } from '@/global/core/chat/utils';
 import { AppTypeEnum } from '@fastgpt/global/core/app/constants';
 async function handler(
   req: NextApiRequest,
