+++ projects/app/src/pages/api/core/chat/team/init.ts
@@ -17,7 +17,7 @@ import { ChatRoleEnum } from '@fastgpt/global/core/chat/constants';
 import { getAppLatestVersion } from '@fastgpt/service/core/app/controller';
 import { FlowNodeTypeEnum } from '@fastgpt/global/core/workflow/node/constant';
 import { AppTypeEnum } from '@fastgpt/global/core/app/constants';
-import transformPreviewHistories from '@/service/core/chat/utils';
+import { transformPreviewHistories } from '@/global/core/chat/utils';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   try {
